requirements.txt
langchain
langgraph
langchain-openai
langchain-community
python-dotenv
pydantic
duckduckgo-search